RCP-227 refactor: dto 폴더구조 변경
